Refactor: extract cats runtime code
